+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added:
 * `CADD` container https://github.com/Clinical-Genomics/BALSAMIC/pull/1222
 * `Container ettiquette` to ReadtheDocs https://github.com/Clinical-Genomics/BALSAMIC/pull/1132
 * `htslib` (samtools, bcftools tabix) container https://github.com/Clinical-Genomics/BALSAMIC/pull/1234
+* Release version support for cache generation https://github.com/Clinical-Genomics/BALSAMIC/pull/1231
 
 Changed:
 ^^^^^^^^